+++ backend/src/test/kotlin/com/book/manager/presentation/controller/BookControllerTest.kt
@@ -2,8 +2,8 @@ package com.book.manager.presentation.controller
 
 import com.book.manager.domain.model.entity.Book
 import com.book.manager.domain.model.id.BookId
-import com.book.manager.presentation.form.BookInfo
-import com.book.manager.presentation.form.GetBookListResponse
+import com.book.manager.presentation.caster.BookInfo
+import com.book.manager.presentation.caster.GetBookListResponse
 import com.book.manager.usecase.BookService
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.registerKotlinModule
